Fix certificate generation layout bug in legacy Edge
This will fix project-user/#127 which occurred in legacy Edge browser.
The problem: The newline \n character that separates the different activities wasn’t appropriately encoded in the url (it was just ommited). This is the reason why there was no new line between the different activities in the generated certificate, it was just ommited…