Query even non joinable courses if user owns course
In the backend, the new default when requesting data from the getCourses route is „return only courses which are still joinable if not otherwise specified via a onlyJoinableCourses“ parameter. This way, when showing all courses in the user area and on the public course page, it will show by default all courses which are joinable, and no other courses. But in the user area in the section with the courses a user has joined (or a user is offering), it should show all the courses, independet of whether they can be still joined or not – because the user then has already joined the course (either as an instructor or as a participant) and thus need to access that course.