Add a public course detail page
This will add a publicly accessible page that displays the details for a course. It will have a header that links to the login/registration form. If a publicly visible course detail page is accessed with a user already logged in, the user is forwarded to the private course detail page, where he can immediately click on the "Teilnehmen" button.