Prevent login page after successful email verification
If someone has successfully verified his email address, the intended behaviour is to immediately redirect that person into the user login (i.e. logging someone in). Before this commit, someone was redirected to the login page and from there into the user section. See PrivateRoute.tsx file (line 29-37). This lead to a short delay and primarily for slow internet connections appeared to stay stuck on the login page (even though in the background it was working to redirect into the user section, which then usually also happens abpruptly from the nowhere). This should be prevented now by setting the authContext’s status appropriately.