Add advanced waiting logic to Login page
Previously if someone has got a login-link (or just manually accesses the login page) there was no way of seeing the background process of logging in. That lead to scenarios where, primarily with slow internet connections, someone accessed the login page and saw the usual „enter your email to get a login token“-UI while the app was trying to login with the provided token in the background (but it wasn’t indicating that). This commit adds some advanced information to the login process and status and presents it to the user.