Fix #72 (time was resetted when creating lecture)
Because moment works with objects and calling a method on a moment object will modify the object itself, this will lead to some hard to debug problems. In this case, we were using the startOf method on the internal date used by the antd-DatePicker. So we have modified the internal representation of the date and reset it to the start of the day, i.e. 00:00. So we were only able to submit dates with a time of 00:00. Cloning the internal date given by antd will fix this problem.