Make university optional during registration
This is a probably only temporary fix (until we have the option to add university manually via entering the name). But this is required since there are universities in foreign countries that aren’t contained in our list, while we still accepting those students. So until we will have a text input, we will need to make this optional, since we only have a select field for universities.