Extend coach registration with jufo cert question
The registration for 1-on-1 project coaches now requests indication of whether the coach has a jufo certificate (that s*he can show to us) or not. It is only asked for coaches that are no university students. And this information is also forwareded to the backend. If the coach has no jufo cert, s*he needs to accept her*his data to be transferred to Stiftung Jugend forscht e.V. to verify their past participation.