Add Berufsschule as possible school type for Jufo
Project coaching (especially the jufo cooperation) should be possible for Berufsschulen as well. Thus adding this option to the registration form for project coaches. This will not add it to registration forms if someone does not register for project coaching. But those tutees/participants are still able to select „Berufsschule“ in their settings.