Change how info on past jufo participation is asked
This commit introduces an open text field that can be used to enter any kind of information, not only year and topic of the past project. This way, it should be easier and less formal to get info on a past jufo participation and, as a consequence, even easier to really verify them.