Fix too small expert detail card in list
If the expert detail card was displayed in the list and the description was too small, then the entire expert card was rendered too small in the middle of the list. That wasn't in line with the rest of the list and those entries with longer descriptions. Therefore this commit will make all those details card the same width – independent of their content.