Hide subjects in setting's information if empty
If no subjects are given, hide the entire section containing the subject information from the settings page's information box. This way the UI is cleaner. Primarily for those that are only participating in project coaching (and thus not providing subjects during registration), hiding that subjects section will clean up the UI.