Fix course filtering by morning/afternoon
It will address and fix corona-school/project-user#236.
This commit will fix the general problem that occurred because of a wrongly implemented date comparison. But it will also change when a course is said to be "in the morning" or "in the afternoon", which will now be according to corona-school/project-user#236.