Add course level allow contact option
This is the frontend part for corona-school/project-user#196, which, on a course level, adds the option for allowing course participants to contact their course instructors.

This might be adjusted to use an array of instructor IDs instead of a single boolean value, depending on https://github.com/corona-school/project-user/issues/195#issuecomment-755855440 and https://github.com/corona-school/project-user/issues/196#issuecomment-755854579.