Add waiting list support
If a course is fully booked, it is now possible to join a waiting list for that course. It will use the new waiting list support of the API to allow joining/leaving a course's waiting list. The button that will allow joining a waiting list will only be visible if the course is full. Otherwise a normal join will be possible. If there is a new empty place in a course, it will allow joining the course while leaving the waiting list at the same time.