Update UI to support contact-correspondent feature
In the settings of a course, a course instructor can now define one of the course's instructors as the so called "correspondent" of the course who will receive, if allowContact for this course is switched on, all contact requests of pupils.
Also see corona-school/backend@2e00ea5662b4a9e8aea88846ba55eb9e12aeb916 and corona-school/project-user#195 and corona-school/project-user#196.