Always allow going to course if it has no next lecture
If a course has no lectures or the next lecture has already began while the lecture after the next lecture is not yet in the near future, it will now be possible to nevertheless start the video chat. Another possible problem existed for the last lecture if it has already started and someone wanna join later. Additionally we need to consider that this is still dependent on the local machine's time.

This way, it is possible to later join a course or restart a course as an instructor in some cases, but it does not fully solve all problems arising from this new limitation.