Add option to manually sort club tag rows
This way, someone can define some club tag IDs for which the corresponding rows of courses displayed in the (public) course overview are shown first. This can be used to focus courses for certain tags that need to be filled with priority. Bad, but quick: Those prioritized club tag IDs need to be manually hardcoded into this code.