Add very simple modal to add guests to course
This actually duplicates the very, very basic modal that is used to invite other instructors to a course. The new modal can be used to invites guests to a course. This commit will also add the corresponding API endpoint request methods. Now an instructor of a course can add up to 5 guests to a course.