Add VSCode devcontainer setup
This commit introduces a basic devcontainer setup for VSCode that should make development and getting started with our codebase as easy as possible. This should create VSCode in a way that is expected to work best with this app, including a prettier setup for automatic code formatting. If the container is started, someone then can start the react app's development mode by running `yarn dev` in a new terminal window.