Do not evaluate tokens while on corona-school.de
For some reason some tokens are evaluated on the corona-school-domain shortly before being redirected to lern-fair.de. The token is then tried to be evaluated on the lern-fair.de-site again which results in an error. This behaviour should be treated more thoroughly. Hence this commit serves only as a quick fix.